+++ src/main/java/org/apiphany/io/ContentType.java
@@ -8,6 +8,7 @@ import java.util.function.Supplier;
 import org.apiphany.ApiMimeType;
 import org.apiphany.lang.Strings;
 import org.morphix.lang.Enums;
+import org.morphix.reflection.Constructors;
 
 /**
  * This enum provides a set of commonly used MIME types and their associated character sets.
@@ -19,102 +20,102 @@ public enum ContentType implements ApiMimeType {
 	/**
 	 * Atom XML content type.
 	 */
-	APPLICATION_ATOM_XML("application/atom+xml", StandardCharsets.UTF_8),
+	APPLICATION_ATOM_XML(Value.APPLICATION_ATOM_XML, StandardCharsets.UTF_8),
 
 	/**
 	 * Form URL-encoded content type.
 	 */
-	APPLICATION_FORM_URLENCODED("application/x-www-form-urlencoded", StandardCharsets.ISO_8859_1),
+	APPLICATION_FORM_URLENCODED(Value.APPLICATION_FORM_URLENCODED, StandardCharsets.ISO_8859_1),
 
 	/**
 	 * JSON content type.
 	 */
-	APPLICATION_JSON("application/json", StandardCharsets.UTF_8),
+	APPLICATION_JSON(Value.APPLICATION_JSON, StandardCharsets.UTF_8),
 
 	/**
 	 * Newline-delimited JSON content type.
 	 */
-	APPLICATION_NDJSON("application/x-ndjson", StandardCharsets.UTF_8),
+	APPLICATION_NDJSON(Value.APPLICATION_NDJSON, StandardCharsets.UTF_8),
 
 	/**
 	 * Binary data (octet-stream) content type.
 	 */
-	APPLICATION_OCTET_STREAM("application/octet-stream"),
+	APPLICATION_OCTET_STREAM(Value.APPLICATION_OCTET_STREAM),
 
 	/**
 	 * PDF content type.
 	 */
-	APPLICATION_PDF("application/pdf", StandardCharsets.UTF_8),
+	APPLICATION_PDF(Value.APPLICATION_PDF, StandardCharsets.UTF_8),
 
 	/**
 	 * SOAP XML content type.
 	 */
-	APPLICATION_SOAP_XML("application/soap+xml", StandardCharsets.UTF_8),
+	APPLICATION_SOAP_XML(Value.APPLICATION_SOAP_XML, StandardCharsets.UTF_8),
 
 	/**
 	 * SVG XML content type.
 	 */
-	APPLICATION_SVG_XML("application/svg+xml", StandardCharsets.UTF_8),
+	APPLICATION_SVG_XML(Value.APPLICATION_SVG_XML, StandardCharsets.UTF_8),
 
 	/**
 	 * XHTML XML content type.
 	 */
-	APPLICATION_XHTML_XML("application/xhtml+xml", StandardCharsets.UTF_8),
+	APPLICATION_XHTML_XML(Value.APPLICATION_XHTML_XML, StandardCharsets.UTF_8),
 
 	/**
 	 * XML content type.
 	 */
-	APPLICATION_XML("application/xml", StandardCharsets.UTF_8),
+	APPLICATION_XML(Value.APPLICATION_XML, StandardCharsets.UTF_8),
 
 	/**
 	 * Problem JSON content type.
 	 */
-	APPLICATION_PROBLEM_JSON("application/problem+json", StandardCharsets.UTF_8),
+	APPLICATION_PROBLEM_JSON(Value.APPLICATION_PROBLEM_JSON, StandardCharsets.UTF_8),
 
 	/**
 	 * Problem XML content type.
 	 */
-	APPLICATION_PROBLEM_XML("application/problem+xml", StandardCharsets.UTF_8),
+	APPLICATION_PROBLEM_XML(Value.APPLICATION_PROBLEM_XML, StandardCharsets.UTF_8),
 
 	/**
 	 * RSS XML content type.
 	 */
-	APPLICATION_RSS_XML("application/rss+xml", StandardCharsets.UTF_8),
+	APPLICATION_RSS_XML(Value.APPLICATION_RSS_XML, StandardCharsets.UTF_8),
 
 	/**
 	 * BMP image content type.
 	 */
-	IMAGE_BMP("image/bmp"),
+	IMAGE_BMP(Value.IMAGE_BMP),
 
 	/**
 	 * GIF image content type.
 	 */
-	IMAGE_GIF("image/gif"),
+	IMAGE_GIF(Value.IMAGE_GIF),
 
 	/**
 	 * JPEG image content type.
 	 */
-	IMAGE_JPEG("image/jpeg"),
+	IMAGE_JPEG(Value.IMAGE_JPEG),
 
 	/**
 	 * PNG image content type.
 	 */
-	IMAGE_PNG("image/png"),
+	IMAGE_PNG(Value.IMAGE_PNG),
 
 	/**
 	 * SVG image content type.
 	 */
-	IMAGE_SVG("image/svg+xml"),
+	IMAGE_SVG(Value.IMAGE_SVG),
 
 	/**
 	 * TIFF image content type.
 	 */
-	IMAGE_TIFF("image/tiff"),
+	IMAGE_TIFF(Value.IMAGE_TIFF),
 
 	/**
 	 * WebP image content type.
 	 */
-	IMAGE_WEBP("image/webp"),
+	IMAGE_WEBP(Value.IMAGE_WEBP),
 
 	/**
 	 * Content type for a full HTTP message, used primarily in the response to a TRACE request. The body contains the exact
@@ -122,52 +123,221 @@ public enum ContentType implements ApiMimeType {
 	 *
 	 * @see <a href="https://tools.ietf.org/html/rfc7230#section-8.3.1">RFC 7230, Section 8.3.1</a>
 	 */
-	MESSAGE_HTTP("message/http"),
+	MESSAGE_HTTP(Value.MESSAGE_HTTP),
 
 	/**
 	 * Multipart form data content type.
 	 */
-	MULTIPART_FORM_DATA("multipart/form-data", StandardCharsets.ISO_8859_1),
+	MULTIPART_FORM_DATA(Value.MULTIPART_FORM_DATA, StandardCharsets.ISO_8859_1),
 
 	/**
 	 * Multipart mixed content type.
 	 */
-	MULTIPART_MIXED("multipart/mixed", StandardCharsets.ISO_8859_1),
+	MULTIPART_MIXED(Value.MULTIPART_MIXED, StandardCharsets.ISO_8859_1),
 
 	/**
 	 * Multipart related content type.
 	 */
-	MULTIPART_RELATED("multipart/related", StandardCharsets.ISO_8859_1),
+	MULTIPART_RELATED(Value.MULTIPART_RELATED, StandardCharsets.ISO_8859_1),
 
 	/**
 	 * HTML text content type.
 	 */
-	TEXT_HTML("text/html", StandardCharsets.UTF_8),
+	TEXT_HTML(Value.TEXT_HTML, StandardCharsets.UTF_8),
 
 	/**
 	 * Markdown text content type.
 	 */
-	TEXT_MARKDOWN("text/markdown", StandardCharsets.UTF_8),
+	TEXT_MARKDOWN(Value.TEXT_MARKDOWN, StandardCharsets.UTF_8),
 
 	/**
 	 * Plain text content type.
 	 */
-	TEXT_PLAIN("text/plain", StandardCharsets.UTF_8),
+	TEXT_PLAIN(Value.TEXT_PLAIN, StandardCharsets.UTF_8),
 
 	/**
 	 * XML text content type.
 	 */
-	TEXT_XML("text/xml", StandardCharsets.UTF_8),
+	TEXT_XML(Value.TEXT_XML, StandardCharsets.UTF_8),
 
 	/**
 	 * Event stream text content type.
 	 */
-	TEXT_EVENT_STREAM("text/event-stream", StandardCharsets.UTF_8),
+	TEXT_EVENT_STREAM(Value.TEXT_EVENT_STREAM, StandardCharsets.UTF_8),
 
 	/**
 	 * Wild card content type (matches any type).
 	 */
-	WILDCARD("*/*");
+	WILDCARD(Value.WILDCARD);
+
+	/**
+	 * Name space for all the {@link String} values used by this enumeration so that they can be used in annotations since
+	 * they need constant expressions.
+	 *
+	 * @author Radu Sebastian LAZIN
+	 */
+	public static class Value {
+
+		/**
+		 * Atom XML content type.
+		 */
+		public static final String APPLICATION_ATOM_XML = "application/atom+xml";
+
+		/**
+		 * Form URL-encoded content type.
+		 */
+		public static final String APPLICATION_FORM_URLENCODED = "application/x-www-form-urlencoded";
+
+		/**
+		 * JSON content type.
+		 */
+		public static final String APPLICATION_JSON = "application/json";
+
+		/**
+		 * Newline-delimited JSON content type.
+		 */
+		public static final String APPLICATION_NDJSON = "application/x-ndjson";
+
+		/**
+		 * Binary data (octet-stream) content type.
+		 */
+		public static final String APPLICATION_OCTET_STREAM = "application/octet-stream";
+
+		/**
+		 * PDF content type.
+		 */
+		public static final String APPLICATION_PDF = "application/pdf";
+
+		/**
+		 * SOAP XML content type.
+		 */
+		public static final String APPLICATION_SOAP_XML = "application/soap+xml";
+
+		/**
+		 * SVG XML content type.
+		 */
+		public static final String APPLICATION_SVG_XML = "application/svg+xml";
+
+		/**
+		 * XHTML XML content type.
+		 */
+		public static final String APPLICATION_XHTML_XML = "application/xhtml+xml";
+
+		/**
+		 * XML content type.
+		 */
+		public static final String APPLICATION_XML = "application/xml";
+
+		/**
+		 * Problem JSON content type.
+		 */
+		public static final String APPLICATION_PROBLEM_JSON = "application/problem+json";
+
+		/**
+		 * Problem XML content type.
+		 */
+		public static final String APPLICATION_PROBLEM_XML = "application/problem+xml";
+
+		/**
+		 * RSS XML content type.
+		 */
+		public static final String APPLICATION_RSS_XML = "application/rss+xml";
+
+		/**
+		 * BMP image content type.
+		 */
+		public static final String IMAGE_BMP = "image/bmp";
+
+		/**
+		 * GIF image content type.
+		 */
+		public static final String IMAGE_GIF = "image/gif";
+
+		/**
+		 * JPEG image content type.
+		 */
+		public static final String IMAGE_JPEG = "image/jpeg";
+
+		/**
+		 * PNG image content type.
+		 */
+		public static final String IMAGE_PNG = "image/png";
+
+		/**
+		 * SVG image content type.
+		 */
+		public static final String IMAGE_SVG = "image/svg+xml";
+
+		/**
+		 * TIFF image content type.
+		 */
+		public static final String IMAGE_TIFF = "image/tiff";
+
+		/**
+		 * WebP image content type.
+		 */
+		public static final String IMAGE_WEBP = "image/webp";
+
+		/**
+		 * Content type for a full HTTP message, used primarily in the response to a TRACE request. The body contains the exact
+		 * HTTP message that was received.
+		 *
+		 * @see <a href="https://tools.ietf.org/html/rfc7230#section-8.3.1">RFC 7230, Section 8.3.1</a>
+		 */
+		public static final String MESSAGE_HTTP = "message/http";
+
+		/**
+		 * Multipart form data content type.
+		 */
+		public static final String MULTIPART_FORM_DATA = "multipart/form-data";
+
+		/**
+		 * Multipart mixed content type.
+		 */
+		public static final String MULTIPART_MIXED = "multipart/mixed";
+
+		/**
+		 * Multipart related content type.
+		 */
+		public static final String MULTIPART_RELATED = "multipart/related";
+
+		/**
+		 * HTML text content type.
+		 */
+		public static final String TEXT_HTML = "text/html";
+
+		/**
+		 * Markdown text content type.
+		 */
+		public static final String TEXT_MARKDOWN = "text/markdown";
+
+		/**
+		 * Plain text content type.
+		 */
+		public static final String TEXT_PLAIN = "text/plain";
+
+		/**
+		 * XML text content type.
+		 */
+		public static final String TEXT_XML = "text/xml";
+
+		/**
+		 * Event stream text content type.
+		 */
+		public static final String TEXT_EVENT_STREAM = "text/event-stream";
+
+		/**
+		 * Wild card content type (matches any type).
+		 */
+		public static final String WILDCARD = "*/*";
+
+		/**
+		 * Hide constructor.
+		 */
+		private Value() {
+			throw Constructors.unsupportedOperationException();
+		}
+	}
 
 	/**
 	 * A map of content type names to their corresponding enum values for easy lookup.